# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NTC Market E-Commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /banners:
    get:
      operationId: getBanners
      tags: [banners]
      summary: Get active promotional banners
      responses:
        "200":
          description: Banners list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Banner"
    post:
      operationId: createBanner
      tags: [banners]
      summary: Create a banner (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BannerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"

  # ─── Coupons ─────────────────────────────────────────────────────────────
...
  /banners/{id}:
    patch:
      operationId: updateBanner
      tags: [banners]
      summary: Update a banner (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BannerInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
    delete:
      operationId: deleteBanner
      tags: [banners]
      summary: Delete a banner (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Wishlist ─────────────────────────────────────────────────────────────
...
  /referrals/validate:
    post:
      operationId: validateReferral
      tags: [referrals]
      summary: Validate a referral code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateReferralInput"
      responses:
        "200":
          description: Referral code is valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferralValidationResult"

  # ─── Delivery slots ───────────────────────────────────────────────────────
...
components:
  schemas:
...
    Banner:
      type: object
      required: [id, title, imageUrl, isActive]
      properties:
        id:
          type: integer
        title:
          type: string
        imageUrl:
          type: string
        linkUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        sortOrder:
          type: integer
    BannerInput:
      type: object
      required: [title, imageUrl]
      properties:
        title:
          type: string
        imageUrl:
          type: string
        linkUrl:
          type: string
        isActive:
          type: boolean
        sortOrder:
          type: integer

    # ─── Coupons ───────────────────────────────────────────────────────────
...
    ValidateReferralInput:
      type: object
      required: [code]
      properties:
        code:
          type: string
    ReferralValidationResult:
      type: object
      required: [valid, referrerName]
      properties:
        valid:
          type: boolean
        referrerName:
          type: string

    # ─── Delivery slots ──────────────────────────────────────────────────────
...